+++ packages/integration-sdk-cli/src/commands/index.ts
@@ -1,5 +1,6 @@
 export * from './collect';
 export * from './compare';
+export * from './diff';
 export * from './visualize';
 export * from './sync';
 export * from './run';